Refactor components
Refactor components, file structure and use of static methods for creating widgets
