+++ Main.py
@@ -1,13 +1,15 @@
 import dash
 import dash_html_components as html
-
-from components.age.age import age_widget
-from components.anestesi.anestesi import anestesi_widget
-from components.asa.asa import asa_widget
-from tab_selection import tab_selection
 import dash_bootstrap_components as dbc
-from components.reset_and_search.Reset_and_search import reset_and_search
-from components.kommuner.kommuner import kommuner_widget
+
+from components.age import Age_widget
+from components.anestesi import Anestesi_widget
+from components.asa import Asa_widget
+from components.Reset_and_search import Reset_and_search
+from components.kommuner import Kommuner_widget
+
+from tab_selection import tab_selection
+
 
 # TODO Kolla upp hur man skulle kunna se värdena på patienten man ska ersätta samtidigt som man letar efter en ny,
 #   för att slippa bläddra fram o tillbaka
@@ -19,17 +21,17 @@ app.layout = html.Div(
     id='Main',
     children=[
         tab_selection(),
-        reset_and_search(),
+        Reset_and_search.reset_and_search(),
 
         # TODO: Move this to proper place:
         html.H1(
             id='h1',
             children='Plando-prototype'),
         dbc.Col([
-            asa_widget(),
-            anestesi_widget(),
-            age_widget(),
-            kommuner_widget(),
+            Asa_widget.asa_widget(),
+            Anestesi_widget.anestesi_widget(),
+            Age_widget.age_widget(),
+            Kommuner_widget.kommuner_widget(),
             ]
         )
